+++ NEWS.md
@@ -1,4 +1,23 @@
-# systemPipeShiny 1.3.15 Dev
+# systemPipeShiny 1.5.10 Dev
+
+## Major Change
+
+-   Redesign of the welcome page. Old content is moved to `about`. Now the welcome page is more clear.
+
+-   Introduce `cookies`, this is used to remember if users have clicked the interactive guide. When they refresh the page, the guide will not pop up again.
+
+-   Adapt SPR to the 2.1.x version.
+
+    -   Add more instructing images to the workflow module.
+    -   A warning message is added if `spsOption("demo", TRUE)`, to let people know some workflow templates will fail if they use the demo server to run jobs.
+
+## Minor Change
+
+-   Fix some text typo, links.
+-   Add more figures as instructions in different modules/tabs.
+-   Text/links fixed in workflow module.
+
+# systemPipeShiny 1.3.15
 
 ## New Feature